Added functionality for IEEE-754
Can now convert to/from IEEE-754 single precision floats. Takes either an integer and bit length, a float, or a bit string. Will enable IEEE 754 mode when length requested or length of binary string is 32, or when passed a float, and will still use the other 4 conversion modes when possible.